Rename Login component file with a capital L to follow naming convention
